Move instant execution cache fingerprint code to `fingerprint` package
