+++ subprojects/instant-execution/src/main/kotlin/org/gradle/instantexecution/DefaultInstantExecution.kt
@@ -18,19 +18,17 @@ package org.gradle.instantexecution
 
 import org.gradle.api.Project
 import org.gradle.api.internal.project.ProjectStateRegistry
-import org.gradle.api.internal.provider.DefaultValueSourceProviderFactory
 import org.gradle.api.internal.provider.ValueSourceProviderFactory
-import org.gradle.api.internal.provider.sources.SystemPropertyValueSource
 import org.gradle.api.invocation.Gradle
 import org.gradle.api.logging.LogLevel
 import org.gradle.api.logging.Logging
 import org.gradle.api.provider.Provider
-import org.gradle.api.provider.ValueSource
-import org.gradle.api.provider.ValueSourceParameters
 import org.gradle.execution.plan.Node
 import org.gradle.initialization.InstantExecution
-import org.gradle.instantexecution.extensions.uncheckedCast
 import org.gradle.instantexecution.extensions.unsafeLazy
+import org.gradle.instantexecution.fingerprint.InstantExecutionCacheInputs
+import org.gradle.instantexecution.fingerprint.InstantExecutionFingerprintChecker
+import org.gradle.instantexecution.fingerprint.InvalidationReason
 import org.gradle.instantexecution.initialization.InstantExecutionStartParameter
 import org.gradle.instantexecution.serialization.DefaultReadContext
 import org.gradle.instantexecution.serialization.DefaultWriteContext
@@ -215,9 +213,8 @@ class DefaultInstantExecution internal constructor(
     private
     suspend fun DefaultWriteContext.encodeFingerprint() {
         withHostIsolate {
-            instantExecutionInputs!!.run {
-                writeCollection(inputFiles)
-                writeCollection(obtainedValues)
+            InstantExecutionFingerprintChecker.FingerprintEncoder.run {
+                encodeFingerprintOf(instantExecutionInputs!!)
             }
         }
     }
@@ -226,72 +223,14 @@ class DefaultInstantExecution internal constructor(
     fun checkFingerprint(): InvalidationReason? =
         withReadContextFor(instantExecutionFingerprintFile) {
             withHostIsolate {
-                checkFingerprintOfInputFiles()
-                    ?: checkFingerprintOfObtainedValues()
-            }
-        }
-
-    private
-    suspend fun DefaultReadContext.checkFingerprintOfInputFiles(): InvalidationReason? {
-        readCollection {
-            val (inputFile, hashCode) = readNonNull<InstantExecutionCacheInputs.InputFile>()
-            if (hashCodeOf(inputFile) != hashCode) {
-                // TODO: log some debug info
-                return "a configuration file has changed"
-            }
-        }
-        return null
-    }
-
-    private
-    suspend fun DefaultReadContext.checkFingerprintOfObtainedValues(): InvalidationReason? {
-        readCollection {
-            val obtainedValue = readNonNull<ObtainedValue>()
-            checkFingerprintValueIsUpToDate(obtainedValue)?.let { reason ->
-                return reason
-            }
-        }
-        return null
-    }
-
-    private
-    fun hashCodeOf(inputFile: File) = virtualFileSystem.hashCodeOf(inputFile)
-
-    private
-    fun checkFingerprintValueIsUpToDate(obtainedValue: ObtainedValue): InvalidationReason? = obtainedValue.run {
-        when (valueSourceType) {
-            SystemPropertyValueSource::class.java -> {
-                // Special case system properties to get them from the host because
-                // this check happens too early in the process, before the system properties
-                // passed in the command line have been propagated.
-                val propertyName = valueSourceParameters
-                    .uncheckedCast<SystemPropertyValueSource.Parameters>()
-                    .propertyName
-                    .get()
-                if (value.get() != System.getProperty(propertyName)) {
-                    "system property '$propertyName' has changed"
-                } else {
-                    null
-                }
-            }
-            else -> {
-                val valueSource = instantiateValueSource()
-                if (value.get() != valueSource.obtain()) {
-                    "a build logic input has changed"
-                } else {
-                    null
+                InstantExecutionFingerprintChecker(
+                    virtualFileSystem,
+                    valueSourceProviderFactory
+                ).run {
+                    checkFingerprint()
                 }
             }
         }
-    }
-
-    private
-    fun ObtainedValue.instantiateValueSource(): ValueSource<Any, ValueSourceParameters> =
-        (valueSourceProviderFactory as DefaultValueSourceProviderFactory).instantiateValueSource(
-            valueSourceType,
-            valueSourceParametersType,
-            valueSourceParameters
-        )
 
     private
     fun attachBuildLogicInputsCollector() {
@@ -564,10 +503,6 @@ class DefaultInstantExecution internal constructor(
     fun systemProperty(propertyName: String) =
         startParameter.systemPropertyArg(propertyName) ?: System.getProperty(propertyName)
 }
-
-
-private
-typealias InvalidationReason = String
 
 
 inline fun <reified T> DefaultInstantExecution.Host.service(): T =
